REF: move copulas into separate modules
