+++ statsmodels/distributions/copula/other_copulas.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Jan 29 19:19:45 2021
+
+Author: Josef Perktold
+License: BSD-3
+
+"""
+
